requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Mail==0.9.1
psycopg2==2.9.9
APScheduler==3.8.1
python-dotenv==0.19.0
Werkzeug==2.0.1
WTForms==2.3.3
SQLAlchemy==1.4.23
pytz==2023.3 